fix resolver shutdown, ensure session is started before we try to close it
It only takes a few ms for the session to start, but if we try to close
the session before it it running then the Close call takes over a full
second due to a timeout on a GOAWAY message that never sends.  So we
just ensure the few ms are spent before we proceed to allow you to
Close (context cancel will still abort immediately).