remove unnecessary session
For solves that didn't require a download attachable we used to create a
special "nil" attachable session for this, but it was unnecessary as we
can use any existing session since we don't care about the download
attachable.  We now just use the first session we find for the current
solve instead of creating and using the special "nil" attachable session.

Also adding a parallel download test that would have caught our need
to rollback #23.